#644 - Policy Rule Exclude conditions should be processed as a logical AND instead of a logical OR

+++ pkg/openapi/validation.go
@@ -238,23 +238,23 @@ func generateEmptyResource(kindSchema *openapi_v2.Schema) interface{} {
 	case "integer":
 		if kindSchema.GetDefault() != nil {
 			val, _ := strconv.Atoi(string(kindSchema.GetDefault().Value.Value))
-			return val
+			return int64(val)
 		}
 		if kindSchema.GetExample() != nil {
 			val, _ := strconv.Atoi(string(kindSchema.GetExample().GetValue().Value))
-			return val
+			return int64(val)
 		}
-		return 0
+		return int64(0)
 	case "number":
 		if kindSchema.GetDefault() != nil {
 			val, _ := strconv.Atoi(string(kindSchema.GetDefault().Value.Value))
-			return val
+			return int64(val)
 		}
 		if kindSchema.GetExample() != nil {
 			val, _ := strconv.Atoi(string(kindSchema.GetExample().GetValue().Value))
-			return val
+			return int64(val)
 		}
-		return 0
+		return int64(0)
 	case "boolean":
 		if kindSchema.GetDefault() != nil {
 			return string(kindSchema.GetDefault().Value.Value) == "true"
